+++ packages/amplify-cli-logger/src/IAmplifyLogger.ts
@@ -1,7 +1,6 @@
 import { LogErrorPayload, LogPayload } from './Types';
 
 export interface IAmplifyLogger {
-  log(message: string): void;
   logError(content: LogErrorPayload): void;
   logInfo(content: LogPayload): void;
   projectLocalLogInit(projecPath: string): void;
