+++ packages/amplify-cli-logger/src/index.ts
@@ -1,6 +1,6 @@
 import { AmplifyLogger } from './AmplifyLogger';
 import { IAmplifyLogger } from './IAmplifyLogger';
 import { constants } from './constants';
-
+export { Redactor } from './Redactor';
 export const logger: IAmplifyLogger = new AmplifyLogger();
 export const LocalLogDirectory = constants.LOG_DIRECTORY;
